Add highscore collection Refactor API and backend
