+++ client/src/api/highscores.js
@@ -1,5 +1,5 @@
 export function addHighscore(highscoreData) {
-  return fetch('/api/questions', {
+  return fetch('/api/highscores', {
     method: 'POST',
     headers: {
       'Content-Type': 'application/json'
