+++ client/src/api/questions.js
@@ -7,3 +7,16 @@ export function addQuestion(questionData) {
     body: JSON.stringify(questionData)
   });
 }
+
+export async function getRandomQuestion(privateCode) {
+  return fetch(`/api/questions/random?privateCode=${privateCode}`, {
+    method: 'GET'
+  })
+    .then(response => {
+      if (response.status !== 200) {
+        throw new Error(response.statusText);
+      }
+      return response;
+    })
+    .then(response => response.json());
+}
