+++ lib/questions.js
@@ -7,25 +7,6 @@ async function addQuestion(questionData) {
   await questionCollection.insertOne({ ...questionData });
 }
 
-async function addHighscore(highscoreData) {
-  const questionCollection = await getCollection(collectionName);
-  await questionCollection.insertOne({ ...highscoreData });
-}
-
-async function getHighscores() {
-  const questionsCollection = await getCollection(collectionName);
-  const cursor = await questionsCollection
-    .find({ score: { $exists: 1 } })
-    .sort({ score: -1 })
-    .limit(10);
-  const data = await cursor.toArray();
-  if (!data) {
-    throw new Error('Can not find question');
-  }
-
-  return data;
-}
-
 async function getRandomQuestion(privateCode) {
   const questionsCollection = await getCollection(collectionName);
   const cursor = await questionsCollection.aggregate([
@@ -48,7 +29,5 @@ async function getRandomQuestion(privateCode) {
   return question;
 }
 
-exports.addHighscore = addHighscore;
 exports.addQuestion = addQuestion;
 exports.getRandomQuestion = getRandomQuestion;
-exports.getHighscores = getHighscores;
